[ar71xx] Cr@ns net config moved to defconfig/
Files in /etc/defconfig/ are copied to /etc/config/ if they don't
already exist at first boot.
Surprisingly, setting a custom network file in /files/etc/config/network
made the AP not boot up (so avoid use this file in the future !)

network file was previously defined in base-files package (and it worked)
but it needed to avoid removal of the file by ar71xx makefile. This
solution has been given up in order to have a better separation of
openwrt files and custom Cr@ns specific configuration.