hostapd_cli: update beacon if ignore_broadcast set
set is a cli command that allows changing a parameter of a interface, for
example :
<< hostapd_cli -p/var/run/hostapd-phy0 -i wlan0-1 set ignore_broadcast_ssid 1

This set "ignore_broadcast_ssid" for bss with iface "wlan0-1" to 1. However,
this command does't modify by itself the effective beacon of the ssid. We still
have to call ieee802_11_set_beacon.

Another solution has been to modify the config file then reload it (via
reload command or SIGHUP), but in this case, hostapd couldn't tell which
exact part of the file changed, and in doubt, had to deauthenticate every
client on every bss, which was very costly for a single bss change.

This patch can be seen as a workaround to this problem, since it allows
changing effectively the ignore_broadcast_ssid without disconnecting
everybody.